+++ Tests/Opc.Ua.Di.Tests/DiLockClientTests.cs
@@ -56,7 +56,7 @@ namespace Opc.Ua.Di.Tests
             SetupCallReturns(sessionMock, new CallMethodResult
             {
                 StatusCode = StatusCodes.Good,
-                OutputArguments = new Variant[] { new Variant((int)0) }
+                OutputArguments = new Variant[] { new Variant(0) }
             }, r => captured = r);
 
             var client = new DiLockClient(
@@ -88,7 +88,7 @@ namespace Opc.Ua.Di.Tests
             SetupCallReturns(sessionMock, new CallMethodResult
             {
                 StatusCode = StatusCodes.Good,
-                OutputArguments = new Variant[] { new Variant((int)1) }
+                OutputArguments = new Variant[] { new Variant(1) }
             }, r => captured = r);
 
             var client = new DiLockClient(
@@ -116,7 +116,7 @@ namespace Opc.Ua.Di.Tests
             SetupCallReturns(sessionMock, new CallMethodResult
             {
                 StatusCode = StatusCodes.Good,
-                OutputArguments = new Variant[] { new Variant((int)0) }
+                OutputArguments = new Variant[] { new Variant(0) }
             }, r => captured = r);
 
             var client = new DiLockClient(
@@ -143,7 +143,7 @@ namespace Opc.Ua.Di.Tests
             SetupCallReturns(sessionMock, new CallMethodResult
             {
                 StatusCode = StatusCodes.Good,
-                OutputArguments = new Variant[] { new Variant((int)0) }
+                OutputArguments = new Variant[] { new Variant(0) }
             }, r => captured = r);
 
             var client = new DiLockClient(
@@ -167,7 +167,7 @@ namespace Opc.Ua.Di.Tests
             SetupCallReturns(sessionMock, new CallMethodResult
             {
                 StatusCode = StatusCodes.BadUserAccessDenied,
-                OutputArguments = new Variant[] { new Variant((int)0) }
+                OutputArguments = new Variant[] { new Variant(0) }
             });
 
             var client = new DiLockClient(
